Code reorganisation to better use config files. Adapted code to use Gemma 3 as local model. Minor package updates
