+++ requirements_aws.txt
@@ -1,7 +1,7 @@
 pandas==2.2.3
-gradio==5.32.0
+gradio==5.34.2
 spaces==0.34.1
-boto3==1.38.5
+boto3==1.38.38
 pyarrow==19.0.1
 openpyxl==3.1.3
 markdown==3.7
@@ -11,8 +11,9 @@ google-generativeai==0.8.4
 html5lib==1.1
 beautifulsoup4==4.12.3
 rapidfuzz==3.10.1
-#llama-cpp-python==0.2.90 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
-llama-cpp-python==0.3.2 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
+llama-cpp-python==0.3.9
+#llama-cpp-python==0.3.2 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
 transformers==4.51.1
+python-dotenv==1.1.0
 numpy==1.26.4
 typing_extensions==4.12.2
