+++ requirements_gpu.txt
@@ -1,7 +1,7 @@
 pandas==2.2.3
-gradio==5.32.0
+gradio==5.34.2
 spaces==0.34.1
-boto3==1.38.5
+boto3==1.38.38
 pyarrow==19.0.1
 openpyxl==3.1.3
 markdown==3.7
@@ -11,11 +11,14 @@ google-generativeai==0.8.4
 html5lib==1.1
 beautifulsoup4==4.12.3
 rapidfuzz==3.10.1
-torch==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu121
+torch==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu124
+llama-cpp-python==0.3.9 -C cmake.args="-DGGML_CUDA=on"
+# If the above doesn't work, try one of the following
 #llama-cpp-python==0.3.4 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
 # Specify exact llama_cpp wheel for huggingface compatibility
 #https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu121/llama_cpp_python-0.3.4-cp311-cp311-linux_x86_64.whl
-https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu121/llama_cpp_python-0.3.4-cp311-cp311-win_amd64.whl # Windows
+#https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu121/llama_cpp_python-0.3.4-cp311-cp311-win_amd64.whl # Windows
 transformers==4.51.1
+python-dotenv==1.1.0
 numpy==1.26.4
 typing_extensions==4.12.2
