+++ tools/dedup_summaries.py
@@ -1,0 +1,602 @@
+import pandas as pd
+from rapidfuzz import process, fuzz
+from typing import List, Tuple
+import re
+import spaces
+import gradio as gr
+from time import time
+from tqdm import tqdm
+
+from tools.prompts import summarise_topic_descriptions_prompt, summarise_topic_descriptions_system_prompt, system_prompt
+from tools.llm_funcs import construct_gemini_generative_model, process_requests, ResponseObject, load_model
+from tools.helper_functions import create_unique_table_df_from_reference_table, load_in_data_file, get_basic_response_data, convert_reference_table_to_pivot_table, wrap_text
+from tools.config import OUTPUT_FOLDER, RUN_LOCAL_MODEL, MAX_COMMENT_CHARS, MAX_TOKENS, TIMEOUT_WAIT, NUMBER_OF_RETRY_ATTEMPTS, MAX_TIME_FOR_LOOP, BATCH_SIZE_DEFAULT, DEDUPLICATION_THRESHOLD, model_name_map
+
+max_tokens = MAX_TOKENS
+timeout_wait = TIMEOUT_WAIT
+number_of_api_retry_attempts = NUMBER_OF_RETRY_ATTEMPTS
+max_time_for_loop = MAX_TIME_FOR_LOOP
+batch_size_default = BATCH_SIZE_DEFAULT
+deduplication_threshold = DEDUPLICATION_THRESHOLD
+max_comment_character_length = MAX_COMMENT_CHARS
+
+# DEDUPLICATION/SUMMARISATION FUNCTIONS
+def deduplicate_categories(category_series: pd.Series, join_series: pd.Series, reference_df: pd.DataFrame, general_topic_series: pd.Series = None, merge_general_topics = "No", merge_sentiment:str="No", threshold: float = 90) -> pd.DataFrame:
+    """
+    Deduplicates similar category names in a pandas Series based on a fuzzy matching threshold,
+    merging smaller topics into larger topics.
+
+    Parameters:
+        category_series (pd.Series): Series containing category names to deduplicate.
+        join_series (pd.Series): Additional series used for joining back to original results.
+        reference_df (pd.DataFrame): DataFrame containing the reference data to count occurrences.
+        threshold (float): Similarity threshold for considering two strings as duplicates.
+
+    Returns:
+        pd.DataFrame: DataFrame with columns ['old_category', 'deduplicated_category'].
+    """
+    # Count occurrences of each category in the reference_df
+    category_counts = reference_df['Subtopic'].value_counts().to_dict()
+
+    # Initialize dictionaries for both category mapping and scores
+    deduplication_map = {}
+    match_scores = {}  # New dictionary to store match scores
+
+    # First pass: Handle exact matches
+    for category in category_series.unique():
+        if category in deduplication_map:
+            continue
+            
+        # Find all exact matches
+        exact_matches = category_series[category_series.str.lower() == category.lower()].index.tolist()
+        if len(exact_matches) > 1:
+            # Find the variant with the highest count
+            match_counts = {match: category_counts.get(category_series[match], 0) for match in exact_matches}
+            most_common = max(match_counts.items(), key=lambda x: x[1])[0]
+            most_common_category = category_series[most_common]
+            
+            # Map all exact matches to the most common variant and store score
+            for match in exact_matches:
+                deduplication_map[category_series[match]] = most_common_category
+                match_scores[category_series[match]] = 100  # Exact matches get score of 100
+
+    # Second pass: Handle fuzzy matches for remaining categories
+    # Create a DataFrame to maintain the relationship between categories and general topics
+    categories_df = pd.DataFrame({
+        'category': category_series,
+        'general_topic': general_topic_series
+    }).drop_duplicates()
+
+    for _, row in categories_df.iterrows():
+        category = row['category']
+        if category in deduplication_map:
+            continue
+
+        current_general_topic = row['general_topic']
+
+        # Filter potential matches to only those within the same General Topic if relevant
+        if merge_general_topics == "No":
+            potential_matches = categories_df[
+                (categories_df['category'] != category) & 
+                (categories_df['general_topic'] == current_general_topic)
+            ]['category'].tolist()
+        else:
+            potential_matches = categories_df[
+                (categories_df['category'] != category)
+            ]['category'].tolist()
+
+        matches = process.extract(category, 
+                                potential_matches,
+                                scorer=fuzz.WRatio,
+                                score_cutoff=threshold)
+
+        if matches:
+            best_match = max(matches, key=lambda x: x[1])
+            match, score, _ = best_match
+
+            if category_counts.get(category, 0) < category_counts.get(match, 0):
+                deduplication_map[category] = match
+                match_scores[category] = score
+            else:
+                deduplication_map[match] = category
+                match_scores[match] = score
+        else:
+            deduplication_map[category] = category
+            match_scores[category] = 100
+
+    # Create the result DataFrame with scores
+    result_df = pd.DataFrame({
+        'old_category': category_series + " | " + join_series,
+        'deduplicated_category': category_series.map(lambda x: deduplication_map.get(x, x)),
+        'match_score': category_series.map(lambda x: match_scores.get(x, 100))  # Add scores column
+    })
+
+    #print(result_df)
+
+    return result_df
+
+def deduplicate_topics(reference_df:pd.DataFrame,
+                       unique_topics_df:pd.DataFrame,
+                       reference_table_file_name:str,
+                       unique_topics_table_file_name:str,
+                       in_excel_sheets:str="",
+                       merge_sentiment:str= "No",
+                       merge_general_topics:str="No",
+                       score_threshold:int=90,
+                       in_data_files:List[str]=[],
+                       chosen_cols:List[str]="",
+                       deduplicate_topics:str="Yes",
+                       output_folder:str=OUTPUT_FOLDER
+                       ):
+    '''
+    Deduplicate topics based on a reference and unique topics table
+    '''
+    output_files = []
+    log_output_files = []
+    file_data = pd.DataFrame()
+
+    reference_table_file_name_no_ext = reference_table_file_name #get_file_name_no_ext(reference_table_file_name)
+    unique_topics_table_file_name_no_ext = unique_topics_table_file_name #get_file_name_no_ext(unique_topics_table_file_name)
+
+    # For checking that data is not lost during the process
+    initial_unique_references = len(reference_df["Response References"].unique())
+
+    if unique_topics_df.empty:
+        unique_topics_df = create_unique_table_df_from_reference_table(reference_df)
+
+        # Then merge the topic numbers back to the original dataframe
+        reference_df = reference_df.merge(
+            unique_topics_df[['General Topic', 'Subtopic', 'Sentiment', 'Topic_number']],
+            on=['General Topic', 'Subtopic', 'Sentiment'],
+            how='left'
+        )     
+
+    if in_data_files and chosen_cols:
+        file_data, data_file_names_textbox, total_number_of_batches = load_in_data_file(in_data_files, chosen_cols, 1, in_excel_sheets)
+    else:
+        out_message = "No file data found, pivot table output will not be created."
+        print(out_message)
+        #raise Exception(out_message)
+
+    # Run through this x times to try to get all duplicate topics
+    if deduplicate_topics == "Yes":
+        for i in range(0, 8):
+            if merge_sentiment == "No":    
+                if merge_general_topics == "No":
+                    reference_df["old_category"] = reference_df["Subtopic"] + " | " + reference_df["Sentiment"]
+                    reference_df_unique = reference_df.drop_duplicates("old_category")
+
+                    deduplicated_topic_map_df = reference_df_unique.groupby(["General Topic", "Sentiment"]).apply(
+                        lambda group: deduplicate_categories(
+                            group["Subtopic"], 
+                            group["Sentiment"], 
+                            reference_df, 
+                            general_topic_series=group["General Topic"],
+                            merge_general_topics="No",
+                            threshold=score_threshold
+                        )
+                    ).reset_index(drop=True)
+                else:
+                    # This case should allow cross-topic matching but is still grouping by Sentiment
+                    reference_df["old_category"] = reference_df["Subtopic"] + " | " + reference_df["Sentiment"]
+                    reference_df_unique = reference_df.drop_duplicates("old_category")
+
+                    deduplicated_topic_map_df = reference_df_unique.groupby("Sentiment").apply(
+                        lambda group: deduplicate_categories(
+                            group["Subtopic"], 
+                            group["Sentiment"], 
+                            reference_df, 
+                            general_topic_series=None,  # Set to None to allow cross-topic matching
+                            merge_general_topics="Yes",
+                            threshold=score_threshold
+                        )
+                    ).reset_index(drop=True)
+            else:
+                if merge_general_topics == "No":
+                    # Update this case to maintain general topic boundaries
+                    reference_df["old_category"] = reference_df["Subtopic"] + " | " + reference_df["Sentiment"]
+                    reference_df_unique = reference_df.drop_duplicates("old_category")
+
+                    deduplicated_topic_map_df = reference_df_unique.groupby("General Topic").apply(
+                        lambda group: deduplicate_categories(
+                            group["Subtopic"], 
+                            group["Sentiment"], 
+                            reference_df, 
+                            general_topic_series=group["General Topic"],
+                            merge_general_topics="No",
+                            merge_sentiment=merge_sentiment, 
+                            threshold=score_threshold
+                        )
+                    ).reset_index(drop=True)
+                else:
+                    # For complete merging across all categories
+                    reference_df["old_category"] = reference_df["Subtopic"] + " | " + reference_df["Sentiment"]
+                    reference_df_unique = reference_df.drop_duplicates("old_category")
+
+                    deduplicated_topic_map_df = deduplicate_categories(
+                        reference_df_unique["Subtopic"], 
+                        reference_df_unique["Sentiment"], 
+                        reference_df, 
+                        general_topic_series=None,  # Set to None to allow cross-topic matching
+                        merge_general_topics="Yes",
+                        merge_sentiment=merge_sentiment,
+                        threshold=score_threshold
+                    ).reset_index(drop=True)
+           
+            if deduplicated_topic_map_df['deduplicated_category'].isnull().all():
+            # Check if 'deduplicated_category' contains any values
+                print("No deduplicated categories found, skipping the following code.")
+
+            else:
+                # Remove rows where 'deduplicated_category' is blank or NaN
+                deduplicated_topic_map_df = deduplicated_topic_map_df.loc[(deduplicated_topic_map_df['deduplicated_category'].str.strip() != '') & ~(deduplicated_topic_map_df['deduplicated_category'].isnull()), ['old_category','deduplicated_category', 'match_score']]
+
+                #deduplicated_topic_map_df.to_csv(output_folder + "deduplicated_topic_map_df_" + str(i) + ".csv", index=None)
+
+                reference_df = reference_df.merge(deduplicated_topic_map_df, on="old_category", how="left")
+
+                reference_df.rename(columns={"Subtopic": "Subtopic_old", "Sentiment": "Sentiment_old"}, inplace=True)
+                # Extract subtopic and sentiment from deduplicated_category
+                reference_df["Subtopic"] = reference_df["deduplicated_category"].str.extract(r'^(.*?) \|')[0]  # Extract subtopic
+                reference_df["Sentiment"] = reference_df["deduplicated_category"].str.extract(r'\| (.*)$')[0]  # Extract sentiment
+
+                # Combine with old values to ensure no data is lost
+                reference_df["Subtopic"] = reference_df["deduplicated_category"].combine_first(reference_df["Subtopic_old"])
+                reference_df["Sentiment"] = reference_df["Sentiment"].combine_first(reference_df["Sentiment_old"])
+
+
+            reference_df.drop(['old_category', 'deduplicated_category', "Subtopic_old", "Sentiment_old"], axis=1, inplace=True, errors="ignore")
+
+            reference_df = reference_df[["Response References", "General Topic", "Subtopic", "Sentiment", "Summary", "Start row of group"]]
+
+            #reference_df["General Topic"] = reference_df["General Topic"].str.lower().str.capitalize() 
+            #reference_df["Subtopic"] = reference_df["Subtopic"].str.lower().str.capitalize() 
+            #reference_df["Sentiment"] = reference_df["Sentiment"].str.lower().str.capitalize() 
+
+            if merge_general_topics == "Yes":
+                # Replace General topic names for each Subtopic with that for the Subtopic with the most responses
+                # Step 1: Count the number of occurrences for each General Topic and Subtopic combination
+                count_df = reference_df.groupby(['Subtopic', 'General Topic']).size().reset_index(name='Count')
+
+                # Step 2: Find the General Topic with the maximum count for each Subtopic
+                max_general_topic = count_df.loc[count_df.groupby('Subtopic')['Count'].idxmax()]
+
+                # Step 3: Map the General Topic back to the original DataFrame
+                reference_df = reference_df.merge(max_general_topic[['Subtopic', 'General Topic']], on='Subtopic', suffixes=('', '_max'), how='left')
+
+                reference_df['General Topic'] = reference_df["General Topic_max"].combine_first(reference_df["General Topic"])        
+
+            if merge_sentiment == "Yes":
+                # Step 1: Count the number of occurrences for each General Topic and Subtopic combination
+                count_df = reference_df.groupby(['Subtopic', 'Sentiment']).size().reset_index(name='Count')
+
+                # Step 2: Determine the number of unique Sentiment values for each Subtopic
+                unique_sentiments = count_df.groupby('Subtopic')['Sentiment'].nunique().reset_index(name='UniqueCount')
+
+                # Step 3: Update Sentiment to 'Mixed' where there is more than one unique sentiment
+                reference_df = reference_df.merge(unique_sentiments, on='Subtopic', how='left')
+                reference_df['Sentiment'] = reference_df.apply(
+                    lambda row: 'Mixed' if row['UniqueCount'] > 1 else row['Sentiment'],
+                    axis=1
+                )
+
+                # Clean up the DataFrame by dropping the UniqueCount column
+                reference_df.drop(columns=['UniqueCount'], inplace=True)
+
+            reference_df = reference_df[["Response References", "General Topic", "Subtopic", "Sentiment", "Summary", "Start row of group"]]
+
+        # Update reference summary column with all summaries
+        reference_df["Summary"] = reference_df.groupby(
+        ["Response References", "General Topic", "Subtopic", "Sentiment"]
+        )["Summary"].transform(' <br> '.join)
+
+        # Check that we have not inadvertantly removed some data during the above process
+        end_unique_references = len(reference_df["Response References"].unique())
+
+        if initial_unique_references != end_unique_references:
+            raise Exception(f"Number of unique references changed during processing: Initial={initial_unique_references}, Final={end_unique_references}")
+        
+        # Drop duplicates in the reference table - each comment should only have the same topic referred to once
+        reference_df.drop_duplicates(['Response References', 'General Topic', 'Subtopic', 'Sentiment'], inplace=True)
+
+
+        # Remake unique_topics_df based on new reference_df
+        unique_topics_df = create_unique_table_df_from_reference_table(reference_df)
+
+        # Then merge the topic numbers back to the original dataframe
+        reference_df = reference_df.merge(
+            unique_topics_df[['General Topic', 'Subtopic', 'Sentiment', 'Topic_number']],
+            on=['General Topic', 'Subtopic', 'Sentiment'],
+            how='left'
+        ) 
+
+        if not file_data.empty:
+            basic_response_data = get_basic_response_data(file_data, chosen_cols)            
+            reference_df_pivot = convert_reference_table_to_pivot_table(reference_df, basic_response_data)
+
+            reference_pivot_file_path = output_folder + reference_table_file_name_no_ext + "_pivot_dedup.csv"
+            reference_df_pivot.to_csv(reference_pivot_file_path, index=None, encoding='utf-8')
+            log_output_files.append(reference_pivot_file_path)
+
+        #reference_table_file_name_no_ext = get_file_name_no_ext(reference_table_file_name)
+        #unique_topics_table_file_name_no_ext = get_file_name_no_ext(unique_topics_table_file_name)
+
+        reference_file_path = output_folder + reference_table_file_name_no_ext + "_dedup.csv"
+        unique_topics_file_path = output_folder + unique_topics_table_file_name_no_ext + "_dedup.csv"
+        reference_df.to_csv(reference_file_path, index = None, encoding='utf-8')
+        unique_topics_df.to_csv(unique_topics_file_path, index=None, encoding='utf-8')
+
+        output_files.append(reference_file_path)
+        output_files.append(unique_topics_file_path)        
+
+        # Outputs for markdown table output
+        unique_table_df_revised_display = unique_topics_df.apply(lambda col: col.map(lambda x: wrap_text(x, max_text_length=500)))
+
+        deduplicated_unique_table_markdown = unique_table_df_revised_display.to_markdown(index=False)
+
+    return reference_df, unique_topics_df, output_files, log_output_files, deduplicated_unique_table_markdown
+
+def sample_reference_table_summaries(reference_df:pd.DataFrame,
+                                     unique_topics_df:pd.DataFrame,
+                                     random_seed:int,
+                                     no_of_sampled_summaries:int=150):
+    
+    '''
+    Sample x number of summaries from which to produce summaries, so that the input token length is not too long.
+    '''
+    
+    all_summaries = pd.DataFrame()
+    output_files = []
+
+    reference_df_grouped = reference_df.groupby(["General Topic", "Subtopic", "Sentiment"])
+
+    if 'Revised summary' in reference_df.columns:
+        out_message = "Summary has already been created for this file"
+        print(out_message)
+        raise Exception(out_message)
+
+    for group_keys, reference_df_group in reference_df_grouped:
+        #print(f"Group: {group_keys}")
+        #print(f"Data: {reference_df_group}")
+
+        if len(reference_df_group["General Topic"]) > 1:
+
+            filtered_reference_df = reference_df_group.reset_index()
+
+            filtered_reference_df_unique = filtered_reference_df.drop_duplicates(["General Topic", "Subtopic", "Sentiment", "Summary"])
+
+            # Sample n of the unique topic summaries. To limit the length of the text going into the summarisation tool
+            filtered_reference_df_unique_sampled = filtered_reference_df_unique.sample(min(no_of_sampled_summaries, len(filtered_reference_df_unique)), random_state=random_seed)
+
+            #topic_summary_table_markdown = filtered_reference_df_unique_sampled.to_markdown(index=False)
+
+            #print(filtered_reference_df_unique_sampled)
+
+            all_summaries = pd.concat([all_summaries, filtered_reference_df_unique_sampled])
+    
+    summarised_references = all_summaries.groupby(["General Topic", "Subtopic", "Sentiment"]).agg({
+    'Response References': 'size',  # Count the number of references
+    'Summary': lambda x: '\n'.join([s.split(': ', 1)[1] for s in x if ': ' in s])  # Join substrings after ': '
+    }).reset_index()
+
+    summarised_references = summarised_references.loc[(summarised_references["Sentiment"] != "Not Mentioned") & (summarised_references["Response References"] > 1)]
+
+    summarised_references_markdown = summarised_references.to_markdown(index=False)
+
+    return summarised_references, summarised_references_markdown, reference_df, unique_topics_df
+
+def summarise_output_topics_query(model_choice:str, in_api_key:str, temperature:float, formatted_summary_prompt:str, summarise_topic_descriptions_system_prompt:str, local_model=[]):
+    conversation_history = []
+    whole_conversation_metadata = []
+
+    # Prepare Gemini models before query       
+    if "gemini" in model_choice:
+        print("Using Gemini model:", model_choice)
+        model, config = construct_gemini_generative_model(in_api_key=in_api_key, temperature=temperature, model_choice=model_choice, system_prompt=system_prompt, max_tokens=max_tokens)
+    else:
+        print("Using AWS Bedrock model:", model_choice)
+        model = model_choice
+        config = {}
+
+    whole_conversation = [summarise_topic_descriptions_system_prompt] 
+
+    # Process requests to large language model
+    responses, conversation_history, whole_conversation, whole_conversation_metadata, response_text = process_requests(formatted_summary_prompt, system_prompt, conversation_history, whole_conversation, whole_conversation_metadata, model, config, model_choice, temperature, local_model=local_model)
+
+    print("Finished summary query")
+
+    if isinstance(responses[-1], ResponseObject):
+        response_texts = [resp.text for resp in responses]
+    elif "choices" in responses[-1]:
+        response_texts = [resp["choices"][0]['text'] for resp in responses]
+    else:
+        response_texts = [resp.text for resp in responses]
+
+    latest_response_text = response_texts[-1]
+
+    #print("latest_response_text:", latest_response_text)
+    #print("Whole conversation metadata:", whole_conversation_metadata)
+
+    return latest_response_text, conversation_history, whole_conversation_metadata
+
+@spaces.GPU
+def summarise_output_topics(summarised_references:pd.DataFrame,
+                            unique_table_df:pd.DataFrame,
+                            reference_table_df:pd.DataFrame,
+                            model_choice:str,
+                            in_api_key:str,
+                            topic_summary_table_markdown:str,
+                            temperature:float,
+                            table_file_name:str,
+                            summarised_outputs:list = [],  
+                            latest_summary_completed:int = 0,
+                            out_metadata_str:str = "",
+                            in_data_files:List[str]=[],
+                            in_excel_sheets:str="",
+                            chosen_cols:List[str]=[],
+                            log_output_files:list[str]=[],
+                            summarise_format_radio:str="Return a summary up to two paragraphs long that includes as much detail as possible from the original text",
+                            output_folder:str=OUTPUT_FOLDER,
+                            output_files:list[str] = [],                            
+                            summarise_topic_descriptions_prompt:str=summarise_topic_descriptions_prompt, summarise_topic_descriptions_system_prompt:str=summarise_topic_descriptions_system_prompt,
+                            do_summaries:str="Yes",                            
+                            progress=gr.Progress(track_tqdm=True)):
+    '''
+    Create better summaries of the raw batch-level summaries created in the first run of the model.
+    '''
+    out_metadata = []
+    local_model = []
+    summarised_output_markdown = ""
+    
+
+    # Check for data for summarisations
+    if not unique_table_df.empty and not reference_table_df.empty:
+        print("Unique table and reference table data found.")
+    else:
+        out_message = "Please upload a unique topic table and reference table file to continue with summarisation."
+        print(out_message)
+        raise Exception(out_message)
+    
+    if 'Revised summary' in reference_table_df.columns:
+        out_message = "Summary has already been created for this file"
+        print(out_message)
+        raise Exception(out_message)
+    
+    # Load in data file and chosen columns if exists to create pivot table later
+    if in_data_files and chosen_cols:
+        file_data, data_file_names_textbox, total_number_of_batches = load_in_data_file(in_data_files, chosen_cols, 1, in_excel_sheets=in_excel_sheets)
+    else:
+        out_message = "No file data found, pivot table output will not be created."
+        print(out_message)
+        raise Exception(out_message)
+   
+    
+    all_summaries = summarised_references["Summary"].tolist()
+    length_all_summaries = len(all_summaries)
+
+    # If all summaries completed, make final outputs
+    if latest_summary_completed >= length_all_summaries:
+        print("All summaries completed. Creating outputs.")
+
+        model_choice_clean = model_name_map[model_choice]   
+        file_name = re.search(r'(.*?)(?:_batch_|_col_)', table_file_name).group(1) if re.search(r'(.*?)(?:_batch_|_col_)', table_file_name) else table_file_name
+        latest_batch_completed = int(re.search(r'batch_(\d+)_', table_file_name).group(1)) if 'batch_' in table_file_name else ""
+        batch_size_number = int(re.search(r'size_(\d+)_', table_file_name).group(1)) if 'size_' in table_file_name else ""
+        in_column_cleaned = re.search(r'col_(.*?)_reference', table_file_name).group(1) if 'col_' in table_file_name else ""
+
+        # Save outputs for each batch. If master file created, label file as master
+        if latest_batch_completed:
+            batch_file_path_details = f"{file_name}_batch_{latest_batch_completed}_size_{batch_size_number}_col_{in_column_cleaned}"
+        else:
+            batch_file_path_details = f"{file_name}_col_{in_column_cleaned}"
+
+        summarised_references["Revised summary"] = summarised_outputs        
+
+        join_cols = ["General Topic", "Subtopic", "Sentiment"]
+        join_plus_summary_cols = ["General Topic", "Subtopic", "Sentiment", "Revised summary"]
+
+        summarised_references_j = summarised_references[join_plus_summary_cols].drop_duplicates(join_plus_summary_cols)
+
+        unique_table_df_revised = unique_table_df.merge(summarised_references_j, on = join_cols, how = "left")
+
+        # If no new summary is available, keep the original
+        unique_table_df_revised["Revised summary"] = unique_table_df_revised["Revised summary"].combine_first(unique_table_df_revised["Summary"])
+
+        unique_table_df_revised = unique_table_df_revised[["General Topic",	"Subtopic",	"Sentiment", "Response References",	"Revised summary"]]        
+
+        reference_table_df_revised = reference_table_df.merge(summarised_references_j, on = join_cols, how = "left")
+        # If no new summary is available, keep the original
+        reference_table_df_revised["Revised summary"] = reference_table_df_revised["Revised summary"].combine_first(reference_table_df_revised["Summary"])
+        reference_table_df_revised = reference_table_df_revised.drop("Summary", axis=1)
+
+        # Remove topics that are tagged as 'Not Mentioned'
+        unique_table_df_revised = unique_table_df_revised.loc[unique_table_df_revised["Sentiment"] != "Not Mentioned", :]
+        reference_table_df_revised = reference_table_df_revised.loc[reference_table_df_revised["Sentiment"] != "Not Mentioned", :]
+        
+          
+            
+
+        if not file_data.empty:
+            basic_response_data = get_basic_response_data(file_data, chosen_cols)
+            reference_table_df_revised_pivot = convert_reference_table_to_pivot_table(reference_table_df_revised, basic_response_data)
+
+            ### Save pivot file to log area
+            reference_table_df_revised_pivot_path = output_folder + batch_file_path_details + "_summarised_reference_table_pivot_" + model_choice_clean + ".csv"
+            reference_table_df_revised_pivot.to_csv(reference_table_df_revised_pivot_path, index=None, encoding='utf-8')
+            log_output_files.append(reference_table_df_revised_pivot_path)
+
+        # Save to file
+        unique_table_df_revised_path = output_folder + batch_file_path_details + "_summarised_unique_topic_table_" + model_choice_clean + ".csv"
+        unique_table_df_revised.to_csv(unique_table_df_revised_path, index = None, encoding='utf-8')
+
+        reference_table_df_revised_path = output_folder + batch_file_path_details + "_summarised_reference_table_" + model_choice_clean + ".csv"
+        reference_table_df_revised.to_csv(reference_table_df_revised_path, index = None, encoding='utf-8')
+
+        output_files.extend([reference_table_df_revised_path, unique_table_df_revised_path])       
+
+        ###
+        unique_table_df_revised_display = unique_table_df_revised.apply(lambda col: col.map(lambda x: wrap_text(x, max_text_length=500)))
+
+        summarised_output_markdown = unique_table_df_revised_display.to_markdown(index=False)
+
+        # Ensure same file name not returned twice
+        output_files = list(set(output_files))
+        log_output_files = list(set(log_output_files))
+
+        return summarised_references, unique_table_df_revised, reference_table_df_revised, output_files, summarised_outputs, latest_summary_completed, out_metadata_str, summarised_output_markdown, log_output_files
+
+    tic = time.perf_counter()
+    
+    #print("Starting with:", latest_summary_completed)
+    #print("Last summary number:", length_all_summaries)
+
+    if (model_choice == "gemma_2b_it_local") & (RUN_LOCAL_MODEL == "1"):
+                progress(0.1, "Loading in Gemma 2b model")
+                local_model, tokenizer = load_model()
+                #print("Local model loaded:", local_model)
+
+    summary_loop_description = "Creating summaries. " + str(latest_summary_completed) + " summaries completed so far."
+    summary_loop = tqdm(range(latest_summary_completed, length_all_summaries), desc="Creating summaries", unit="summaries")   
+
+    if do_summaries == "Yes":
+        for summary_no in summary_loop:
+
+            print("Current summary number is:", summary_no)
+
+            summary_text = all_summaries[summary_no]
+            #print("summary_text:", summary_text)
+            formatted_summary_prompt = [summarise_topic_descriptions_prompt.format(summaries=summary_text, summary_format=summarise_format_radio)]
+
+            try:
+                response, conversation_history, metadata = summarise_output_topics_query(model_choice, in_api_key, temperature, formatted_summary_prompt, summarise_topic_descriptions_system_prompt, local_model)
+                summarised_output = response
+                summarised_output = re.sub(r'\n{2,}', '\n', summarised_output)  # Replace multiple line breaks with a single line break
+                summarised_output = re.sub(r'^\n{1,}', '', summarised_output)  # Remove one or more line breaks at the start
+                summarised_output = summarised_output.strip()
+            except Exception as e:
+                print(e)
+                summarised_output = ""
+
+            summarised_outputs.append(summarised_output)
+            out_metadata.extend(metadata)
+            out_metadata_str = '. '.join(out_metadata)
+
+            latest_summary_completed += 1
+
+            # Check if beyond max time allowed for processing and break if necessary
+            toc = time.perf_counter()
+            time_taken = tic - toc
+
+            if time_taken > max_time_for_loop:
+                print("Time taken for loop is greater than maximum time allowed. Exiting and restarting loop")
+                summary_loop.close()
+                tqdm._instances.clear()
+                break
+
+    # If all summaries completeed
+    if latest_summary_completed >= length_all_summaries:
+        print("At last summary.")
+
+    output_files = list(set(output_files))
+
+    return summarised_references, unique_table_df, reference_table_df, output_files, summarised_outputs, latest_summary_completed, out_metadata_str, summarised_output_markdown, log_output_files
